feat(unityengine): fix labels and some d3d11on12 games maybe

+++ src/games/unityengine/tonemap.hlsl
@@ -293,7 +293,7 @@ float3 applyUserTonemapACES(float3 untonemapped, int curve = 0, float DSG = 0.98
                                                                      : renodx::tonemap::config::hue_correction_type::CUSTOM;
   config.hue_correction_strength = injectedData.toneMapHueCorrection;
   config.hue_correction_color = lerp(untonemapped, hueCorrectionColor, injectedData.toneMapHueShift);
-  config.reno_drt_hue_correction_method = (int)injectedData.toneMapHueProcessor;
+  config.reno_drt_hue_correction_method = injectedData.toneMapHueProcessor;
   config.reno_drt_tone_map_method = injectedData.toneMapType == 4.f ? renodx::tonemap::renodrt::config::tone_map_method::REINHARD
                                                                     : renodx::tonemap::renodrt::config::tone_map_method::DANIELE;
   config.reno_drt_working_color_space = (int)injectedData.toneMapColorSpace;
@@ -353,7 +353,7 @@ float3 applyUserTonemapNeutral(float3 untonemapped, float peak = 0) {
                                                                      : renodx::tonemap::config::hue_correction_type::CUSTOM;
   config.hue_correction_strength = injectedData.toneMapHueCorrection;
   config.hue_correction_color = lerp(untonemapped, hueCorrectionColor, injectedData.toneMapHueShift);
-  config.reno_drt_hue_correction_method = (int)injectedData.toneMapHueProcessor;
+  config.reno_drt_hue_correction_method = injectedData.toneMapHueProcessor;
   config.reno_drt_tone_map_method = injectedData.toneMapType - 2.f;
   config.reno_drt_per_channel = injectedData.toneMapPerChannel != 0.f;
   config.reno_drt_working_color_space = 0;
@@ -387,7 +387,7 @@ float3 applyUserNoTonemap(float3 untonemapped) {
   config.hue_correction_strength = injectedData.toneMapHueCorrection;
   config.hue_correction_color = lerp(untonemapped, renodx::tonemap::renodrt::NeutralSDR(untonemapped, true), injectedData.toneMapHueShift);
   config.reno_drt_tone_map_method = renodx::tonemap::renodrt::config::tone_map_method::HERMITE_SPLINE;
-  config.reno_drt_hue_correction_method = (int)injectedData.toneMapHueProcessor;
+  config.reno_drt_hue_correction_method = injectedData.toneMapHueProcessor;
   config.reno_drt_blowout = 1.f - injectedData.colorGradeBlowout;
   config.reno_drt_per_channel = injectedData.toneMapPerChannel != 0.f;
   config.reno_drt_working_color_space = 0;
@@ -440,7 +440,7 @@ float3 applyUserTonemapSapphire(float3 untonemapped) {
                                                                      : renodx::tonemap::config::hue_correction_type::CUSTOM;
   config.hue_correction_strength = injectedData.toneMapHueCorrection;
   config.hue_correction_color = lerp(untonemapped, hueCorrectionColor, injectedData.toneMapHueShift);
-  config.reno_drt_hue_correction_method = (int)injectedData.toneMapHueProcessor;
+  config.reno_drt_hue_correction_method = injectedData.toneMapHueProcessor;
   config.reno_drt_tone_map_method = injectedData.toneMapType == 4.f ? renodx::tonemap::renodrt::config::tone_map_method::REINHARD
                                                                     : renodx::tonemap::renodrt::config::tone_map_method::DANIELE;
   config.reno_drt_per_channel = injectedData.toneMapPerChannel != 0.f;
